feat: make the factor enrollment resilient to the CP Database outage
related-to AM-3476
