+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-core/src/main/java/io/gravitee/am/gateway/handler/root/resources/handler/webauthn/WebAuthnLoginHandler.java
@@ -21,6 +21,7 @@ import io.gravitee.am.gateway.handler.common.auth.user.UserAuthenticationManager
 import io.gravitee.am.gateway.handler.common.factor.FactorManager;
 import io.gravitee.am.gateway.handler.common.utils.Tuple;
 import io.gravitee.am.gateway.handler.common.vertx.web.auth.user.User;
+import io.gravitee.am.gateway.handler.root.service.user.UserService;
 import io.gravitee.am.identityprovider.api.AuthenticationContext;
 import io.gravitee.am.model.Credential;
 import io.gravitee.am.model.Domain;
@@ -28,7 +29,6 @@ import io.gravitee.am.model.ReferenceType;
 import io.gravitee.am.model.login.WebAuthnSettings;
 import io.gravitee.am.model.oidc.Client;
 import io.gravitee.am.service.CredentialService;
-import io.gravitee.am.service.FactorService;
 import io.gravitee.common.http.HttpHeaders;
 import io.gravitee.common.http.MediaType;
 import io.reactivex.rxjava3.core.Completable;
@@ -63,13 +63,13 @@ public class WebAuthnLoginHandler extends WebAuthnHandler {
     private final WebAuthn webAuthn;
     private final String origin;
 
-    public WebAuthnLoginHandler(FactorService factorService,
+    public WebAuthnLoginHandler(UserService userService,
                                 FactorManager factorManager,
                                 Domain domain,
                                 WebAuthn webAuthn,
                                 CredentialService credentialService,
                                 UserAuthenticationManager userAuthenticationManager) {
-        setFactorService(factorService);
+        setUserService(userService);
         setFactorManager(factorManager);
         setCredentialService(credentialService);
         setUserAuthenticationManager(userAuthenticationManager);
