+++ gravitee-am-service/src/main/java/io/gravitee/am/service/UserService.java
@@ -19,7 +19,6 @@ import io.gravitee.am.model.ReferenceType;
 import io.gravitee.am.model.User;
 import io.gravitee.am.model.analytics.AnalyticsQuery;
 import io.gravitee.am.model.common.Page;
-import io.gravitee.am.model.factor.EnrolledFactor;
 import io.gravitee.am.repository.management.api.CommonUserRepository.UpdateActions;
 import io.gravitee.am.service.model.NewUser;
 import io.gravitee.am.service.model.UpdateUser;
@@ -60,8 +59,6 @@ public interface UserService extends CommonUserService {
     Single<Map<Object, Object>> statistics(AnalyticsQuery query);
 
     Completable deleteByDomain(String domainId);
-
-    Single<User> upsertFactor(String userId, EnrolledFactor enrolledFactor, io.gravitee.am.identityprovider.api.User principal);
 
     Completable removeFactor(String userId, String factorId, io.gravitee.am.identityprovider.api.User principal);
