+++ gravitee-am-service/src/main/java/io/gravitee/am/service/impl/FactorServiceImpl.java
@@ -24,7 +24,6 @@ import io.gravitee.am.model.Factor;
 import io.gravitee.am.model.ReferenceType;
 import io.gravitee.am.model.common.event.Event;
 import io.gravitee.am.model.common.event.Payload;
-import io.gravitee.am.model.factor.EnrolledFactor;
 import io.gravitee.am.repository.management.api.FactorRepository;
 import io.gravitee.am.service.ApplicationService;
 import io.gravitee.am.service.AuditService;
@@ -229,16 +228,5 @@ public class FactorServiceImpl implements FactorService {
                     return Completable.error(new TechnicalManagementException(
                             String.format("An error occurs while trying to delete factor: %s", factorId), ex));
                 });
-    }
-
-    @Override
-    public Single<io.gravitee.am.model.User> enrollFactor(String userId, EnrolledFactor enrolledFactor, User principal) {
-        LOGGER.debug("Enrolling factor {}", enrolledFactor.getFactorId());
-        try {
-            return userService.upsertFactor(userId, enrolledFactor, principal);
-        } catch (Exception ex) {
-            LOGGER.error("An error occurs while enrolling factor: {}", enrolledFactor.getFactorId(), ex);
-            return Single.error(ex);
-        }
     }
 }
