+++ gravitee-am-service/src/main/java/io/gravitee/am/service/impl/UserServiceImpl.java
@@ -16,15 +16,11 @@
 package io.gravitee.am.service.impl;
 
 import io.gravitee.am.common.audit.EventType;
-import io.gravitee.am.common.exception.mfa.InvalidFactorAttributeException;
 import io.gravitee.am.model.ReferenceType;
 import io.gravitee.am.model.User;
 import io.gravitee.am.model.analytics.AnalyticsQuery;
 import io.gravitee.am.model.common.Page;
 import io.gravitee.am.model.factor.EnrolledFactor;
-import io.gravitee.am.model.factor.EnrolledFactorChannel;
-import io.gravitee.am.model.factor.FactorStatus;
-import io.gravitee.am.model.scim.Attribute;
 import io.gravitee.am.repository.management.api.CommonUserRepository.UpdateActions;
 import io.gravitee.am.repository.management.api.UserRepository;
 import io.gravitee.am.service.AuditService;
@@ -46,15 +42,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
-import static com.google.common.base.Strings.isNullOrEmpty;
-import static java.util.Optional.ofNullable;
 
 /**
  * @author David BRASSELY (david.brassely at graviteesource.com)
@@ -229,97 +220,6 @@ public class UserServiceImpl extends AbstractUserService implements UserService 
     }
 
     @Override
-    public Single<User> upsertFactor(String userId, EnrolledFactor enrolledFactor, io.gravitee.am.identityprovider.api.User principal) {
-        return findById(userId)
-                .switchIfEmpty(Single.error(new UserNotFoundException(userId)))
-                .flatMap(oldUser -> {
-                    User user = new User(oldUser);
-                    List<EnrolledFactor> enrolledFactors = user.getFactors();
-                    if (enrolledFactors == null || enrolledFactors.isEmpty()) {
-                        enrolledFactors = Collections.singletonList(enrolledFactor);
-                    } else {
-                        // if current factor is primary, set the others to secondary
-                        if (Boolean.TRUE.equals(enrolledFactor.isPrimary())) {
-                            enrolledFactors.forEach(e -> e.setPrimary(false));
-                        }
-                        // if the Factor already exists, update the target and the security value
-                        var foundFactor = enrolledFactors.stream()
-                                .filter(existingFactor -> existingFactor.getFactorId().equals(enrolledFactor.getFactorId()))
-                                .findFirst();
-                        if (foundFactor.isPresent()) {
-                            var factorToUpdate = new EnrolledFactor(foundFactor.get());
-                            factorToUpdate.setStatus(ofNullable(enrolledFactor.getStatus()).orElse(factorToUpdate.getStatus()));
-                            factorToUpdate.setChannel(ofNullable(enrolledFactor.getChannel()).orElse(factorToUpdate.getChannel()));
-                            factorToUpdate.setSecurity(ofNullable(enrolledFactor.getSecurity()).orElse(factorToUpdate.getSecurity()));
-                            factorToUpdate.setPrimary(ofNullable(enrolledFactor.isPrimary()).orElse(factorToUpdate.isPrimary()));
-                            // update the factor
-                            enrolledFactors.removeIf(ef -> factorToUpdate.getFactorId().equals(ef.getFactorId()));
-                            enrolledFactors.add(factorToUpdate);
-                        } else {
-                            enrolledFactors.add(enrolledFactor);
-                        }
-                    }
-                    user.setFactors(enrolledFactors);
-
-                    if (enrolledFactor.getChannel() != null && EnrolledFactorChannel.Type.SMS.equals(enrolledFactor.getChannel().getType())) {
-                        // MFA SMS currently used, preserve the phone number into the user profile if not yet present
-                        List<Attribute> phoneNumbers = user.getPhoneNumbers();
-                        if (phoneNumbers == null) {
-                            phoneNumbers = new ArrayList<>();
-                            user.setPhoneNumbers(phoneNumbers);
-                        }
-                        String enrolledPhoneNumber = enrolledFactor.getChannel().getTarget();
-                        if (isNullOrEmpty(enrolledPhoneNumber)) {
-                            return Single.error(new InvalidFactorAttributeException("Phone Number required to enroll SMS factor"));
-                        }
-                        if (phoneNumbers.stream().noneMatch(p -> p.getValue().equals(enrolledPhoneNumber))) {
-                            Attribute newPhoneNumber = new Attribute();
-                            newPhoneNumber.setType("mobile");
-                            newPhoneNumber.setPrimary(phoneNumbers.isEmpty());
-                            newPhoneNumber.setValue(enrolledPhoneNumber);
-                            phoneNumbers.add(newPhoneNumber);
-                        }
-                    }
-                    if (enrolledFactor.getChannel() != null && EnrolledFactorChannel.Type.EMAIL.equals(enrolledFactor.getChannel().getType())) {
-                        // MFA EMAIL currently used, preserve the email into the user profile if not yet present
-                        String email = user.getEmail();
-                        String enrolledEmail = enrolledFactor.getChannel().getTarget();
-                        if (isNullOrEmpty(enrolledEmail)) {
-                            return Single.error(new InvalidFactorAttributeException("Email address required to enroll Email factor"));
-                        }
-                        if (email == null) {
-                            user.setEmail(enrolledEmail);
-                        } else if (!email.equals(enrolledEmail)){
-                            // an email is already present but doesn't match the one provided as security factor
-                            // register this email in the user profile.
-                            List<Attribute> emails = user.getEmails();
-                            if (emails == null) {
-                                emails = new ArrayList<>();
-                                user.setEmails(emails);
-                            }
-                            if (emails.stream().noneMatch(p -> p.getValue().equals(enrolledEmail))) {
-                                Attribute additionalEmail = new Attribute();
-                                additionalEmail.setPrimary(false);
-                                additionalEmail.setValue(enrolledEmail);
-                                emails.add(additionalEmail);
-                            }
-                        }
-                    }
-                    return update(user)
-                            .doOnSuccess(user1 -> LOGGER.debug("Factor {} upserted for user {}", enrolledFactor.getFactorId(), user1.getId()))
-                            .doOnSuccess(user1 -> {
-                               if (needToAuditUserFactorsOperation(user1, oldUser)) {
-                                    // remove sensitive data about factors
-                                    removeSensitiveFactorsData(user1.getFactors());
-                                    removeSensitiveFactorsData(oldUser.getFactors());
-                                    auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type(EventType.USER_UPDATED).user(user1).oldValue(oldUser));
-                                }
-                            })
-                            .doOnError(throwable -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type(EventType.USER_UPDATED).throwable(throwable)));
-                });
-    }
-
-    @Override
     public Completable removeFactor(String userId, String factorId, io.gravitee.am.identityprovider.api.User principal) {
         return findById(userId)
                 .switchIfEmpty(Maybe.error(new UserNotFoundException(userId)))
@@ -338,67 +238,5 @@ public class UserServiceImpl extends AbstractUserService implements UserService 
                             .doOnError(throwable -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type(EventType.USER_UPDATED).throwable(throwable)))
                             .ignoreElement();
                 });
-    }
-
-    private void removeSensitiveFactorsData(List<EnrolledFactor> enrolledFactors) {
-        if (enrolledFactors == null) {
-            return;
-        }
-        enrolledFactors
-                .forEach(enrolledFactor -> enrolledFactor.setSecurity(null));
-    }
-
-    private boolean needToAuditUserFactorsOperation(User newUser, User oldUser) {
-        final List<EnrolledFactor> newEnrolledFactors = newUser.getFactors() != null ? newUser.getFactors() : Collections.emptyList();
-        final List<EnrolledFactor> oldEnrolledFactors = oldUser.getFactors() != null ? oldUser.getFactors() : Collections.emptyList();
-
-        // if new enrolled factor, create an audit
-        if (newEnrolledFactors.size() != oldEnrolledFactors.size()) {
-            return true;
-        }
-
-        // if enrolled factors not match, create an audit
-        return newEnrolledFactors
-                .stream()
-                .anyMatch(newEnrolledFactor -> oldEnrolledFactors
-                        .stream()
-                        .anyMatch(oldEnrolledFactor -> {
-                            if (!newEnrolledFactor.getFactorId().equals(oldEnrolledFactor.getFactorId())) {
-                                return false;
-                            }
-                            // check if enrolled factor was in pending activation
-                            if (oldEnrolledFactor.getStatus().equals(FactorStatus.PENDING_ACTIVATION)) {
-                                return true;
-                            }
-                            if (oldEnrolledFactor.getChannel() != null) {
-                                // check if email has changed
-                                if (EnrolledFactorChannel.Type.EMAIL.equals(oldEnrolledFactor.getChannel().getType())) {
-                                    return emailInformationHasChanged(newUser, oldUser);
-                                }
-                                // check if phoneNumber has changed
-                                if (EnrolledFactorChannel.Type.SMS.equals(oldEnrolledFactor.getChannel().getType())) {
-                                    return phoneNumberInformationHasChanged(newUser, oldUser);
-                                }
-                            }
-                            return false;
-                        }));
-    }
-
-    private boolean emailInformationHasChanged(User newUser, User oldUser) {
-        if (oldUser.getEmail() == null
-                && newUser.getEmail() != null) {
-            return true;
-        }
-
-        // if email list not match, create an audit
-        final List<Attribute> newEmails = ofNullable(newUser.getEmails()).orElse(Collections.emptyList());
-        final List<Attribute> oldEmails = ofNullable(oldUser.getEmails()).orElse(Collections.emptyList());
-        return newEmails.size() != oldEmails.size();
-    }
-
-    private boolean phoneNumberInformationHasChanged(User newUser, User oldUser) {
-        final List<Attribute> newPhoneNumbers = ofNullable(newUser.getPhoneNumbers()).orElse(Collections.emptyList());
-        final List<Attribute> oldPhoneNumbers = ofNullable(oldUser.getPhoneNumbers()).orElse(Collections.emptyList());
-        return newPhoneNumbers.size() != oldPhoneNumbers.size();
     }
 }
